requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
appdirs==1.4.4
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
cryptography==37.0.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
lxml==4.9.1
mysql-connector-python==8.0.29
outcome==1.2.0
packaging==21.3
pluggy==1.0.0
protobuf==4.21.2
py==1.11.0
pybrowsers==0.4.1
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.1
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tomli==2.0.1
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.9
webdriver-manager==3.8.0
webdrivermanager==0.10.0
wsproto==1.1.0
